+++ docs/api/structures/input-event.md
@@ -1,6 +1,6 @@
 # InputEvent Object
 
-* `modifiers` String[] - An array of modifiers of the event, can
-  be `shift`, `control`, `alt`, `meta`, `isKeypad`, `isAutoRepeat`,
-  `leftButtonDown`, `middleButtonDown`, `rightButtonDown`, `capsLock`,
-  `numLock`, `left`, `right`.
+* `modifiers` String[] (optional) - An array of modifiers of the event, can
+  be `shift`, `control`, `ctrl`, `alt`, `meta`, `command`, `cmd`, `isKeypad`,
+  `isAutoRepeat`, `leftButtonDown`, `middleButtonDown`, `rightButtonDown`,
+  `capsLock`, `numLock`, `left`, `right`.
