+++ script/codesign/import-testing-cert-ci.sh
@@ -7,8 +7,6 @@ security create-keychain -p $KEYCHAIN_PASSWORD $KEY_CHAIN
 security default-keychain -s $KEY_CHAIN
 # Unlock the keychain
 security unlock-keychain -p $KEYCHAIN_PASSWORD $KEY_CHAIN
-# Set keychain locking timeout to 3600 seconds
-security set-keychain-settings -t 3600 -u $KEY_CHAIN
 
 # Add certificates to keychain and allow codesign to access them
 security import "$(dirname $0)"/signing.cer -k $KEY_CHAIN -A /usr/bin/codesign
@@ -16,7 +14,7 @@ security import "$(dirname $0)"/signing.pem -k $KEY_CHAIN -A /usr/bin/codesign
 security import "$(dirname $0)"/signing.p12 -k $KEY_CHAIN -P $SPEC_KEY_PASSWORD -A /usr/bin/codesign
 
 echo "Add keychain to keychain-list"
-security list-keychains -s mac-build.keychain
+security list-keychains -s $KEY_CHAIN
 
 echo "Setting key partition list"
 security set-key-partition-list -S apple-tool:,apple: -s -k $KEYCHAIN_PASSWORD $KEY_CHAIN
