+++ spec-main/ambient.d.ts
@@ -16,6 +16,7 @@ declare namespace Electron {
 
   interface WebContents {
     getOwnerBrowserWindow(): BrowserWindow;
+    getWebPreferences(): any;
   }
 
   interface Session {
